thumbnail cascade is flaky
